- relacionamento ManyToMany entre user e project, utilizando a pivot table project_members - criação teste de middleware de permissão de visualização - criação de métodos de permissão em ProjectService e ProjectController

+++ app/Http/Controllers/ProjectController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 
 use FRD\Http\Requests;
 use FRD\Http\Controllers\Controller;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 
 class ProjectController extends Controller
 {
@@ -27,7 +28,7 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return $this->projectRepository->with(['owner','client'])->all();
+        return $this->projectRepository->with(['owner','client'])->findWhere(['owner_id' => Authorizer::getResourceOwnerId()]);
     }
 
     /**
@@ -49,7 +50,12 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
+        if (!$this->checkPermissions($id)) {
+            return ['Error' => 'Access forbidden'];
+        }
+
         return $this->projectService->find($id);
+
     }
 
     /**
@@ -61,6 +67,10 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (!$this->checkProjectOwner($id)) {
+            return ['Error' => 'Access forbidden'];
+        }
+
         return $this->projectService->update($request->all(), $id);
     }
 
@@ -72,12 +82,47 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if (!$this->checkProjectOwner($id)) {
+            return ['Error' => 'Access forbidden'];
+        }
+
         return $this->projectService->destroy($id);
     }
 
     public function members($id)
     {
         return $this->projectService->find($id)->project_members;
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        $userID = Authorizer::getResourceOwnerId();
+
+        if (!$this->projectService->isOwner($projectId, $userID)) {
+            return false;
+        }
+
+        return true;
+    }
+
+    private function checkProjectMember($projectId)
+    {
+        $userID = Authorizer::getResourceOwnerId();
+
+        if (!$this->projectService->isMember($projectId, $userID)) {
+            return false;
+        }
+
+        return true;
+    }
+
+    public function checkPermissions($projectId)
+    {
+        if ($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId)) {
+            return true;
+        }
+
+        return false;
     }
 
 }
